refactor for relay too
